bpf: Add helpers to dequeue from a PIFO map
This adds a new helper to dequeue a packet from a PIFO map,
bpf_packet_dequeue(). The helper returns a refcounted pointer to the packet
dequeued from the map; the reference must be released either by dropping
the packet, or by returning it to the caller. This is achieved through
bpf_packet_drop() and bpf_packet_return() respectively.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>