Added xdp_scheduler_tester_user
This commit adds the xdp_scheduler_tester program and its companion
trace file called xdp_pifo.trace.

The xdp_scheduler_tester is a program to validate XDP schedulers that
use the new dequeue BPF hook. It parses trace files that create and
enqueue simulated packets to XDP, which we dequeue to verify if the
packets match the expected dequeuing order. An example FIFO scheduler
example would be as follows:
udp dst port=8080 ip=::1
udp dst port=8081 ip=::2
dequeue udp dst port=8080 ip=::1
dequeue udp dst port=8081 ip=::2

The xdp_scheduler_tester program has a simple line-based parser that
uses a lookup table of structs to see which function to call. These
structs make it easy to extend the syntax without changing the internal
functions. This commit only provides a handful of commands.

The known issues in this commit are that the parser is strict on the
order of subsection arguments when there are more than one, which it
should not be. It does not check if you are running the program with
root privileges. And this version does not allow you to load external
BPF ELF binaries.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>