Added a strict priority PIFO with a bug
This commit has an issue with the BPF code not finding the IPPROTO_UDP
type in the BPF code in xdp_scheduler_pifo.bpf.c. I am not sure if it is
an issue related to handling the packet in the userspace program or if I
am doing something wrong in the BPF code.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>